+++ baybe/recommender.py
@@ -6,19 +6,20 @@ Recommender classes for optimizing acquisition functions.
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Dict, Optional, Type
+from typing import ClassVar, Dict, Optional, Type
 
 import pandas as pd
 import torch
 from botorch.acquisition import AcquisitionFunction
-from botorch.optim import optimize_acqf_discrete
+from botorch.optim import optimize_acqf, optimize_acqf_discrete
 
-from .utils import isabstract, to_tensor
-
-
-class NoMCAcquisitionFunctionError(Exception):
-    """An exception raised when a Monte Carlo acquisition function is required
-    but an analytical acquisition function has been selected by the user."""
+from .searchspace import SearchSpace
+from .utils import (
+    IncompatibleSearchSpaceError,
+    isabstract,
+    NoMCAcquisitionFunctionError,
+    to_tensor,
+)
 
 
 class Recommender(ABC):
@@ -32,6 +33,9 @@ class Recommender(ABC):
     type: str
     SUBCLASSES: Dict[str, Type[Recommender]] = {}
 
+    compatible_discrete: ClassVar[bool]
+    compatible_continuous: ClassVar[bool]
+
     def __init__(self, acquisition_function: Optional[AcquisitionFunction]):
         self.acquisition_function = acquisition_function
 
@@ -42,22 +46,88 @@ class Recommender(ABC):
         if not isabstract(cls):
             cls.SUBCLASSES[cls.type] = cls
 
-    @abstractmethod
-    def recommend(self, candidates: pd.DataFrame, batch_quantity: int = 1) -> pd.Index:
+    def recommend(
+        self,
+        searchspace: SearchSpace,
+        batch_quantity: int = 1,
+        allow_repeated_recommendations: bool = False,
+        allow_recommending_already_measured: bool = True,
+    ) -> pd.DataFrame:
         """
         Recommends the next experiments to be conducted.
 
         Parameters
         ----------
-        candidates : pd.DataFrame
-            The features of all candidate experiments that could be conducted next.
+        searchspace : SearchSpace
+            The search space from which recommendations should be provided.
         batch_quantity : int
             The number of experiments to be conducted in parallel.
+        allow_repeated_recommendations : bool
+            Whether points whose discrete parts were already recommended can be
+            recommended again.
+        allow_recommending_already_measured : bool
+            Whether points whose discrete parts were already measured can be
+            recommended again.
 
         Returns
         -------
-        The DataFrame indices of the specific experiments selected.
+        The DataFrame with the specific experiments recommended.
         """
+        # TODO[11179]: Potentially move call to get_candidates from _recommend to here
+        # TODO[11179]: Potentially move metadata update from _recommend to here
+
+        # Validate the search space
+        self.check_searchspace_compatibility(searchspace)
+
+        return self._recommend(
+            searchspace,
+            batch_quantity,
+            allow_repeated_recommendations,
+            allow_recommending_already_measured,
+        )
+
+    @abstractmethod
+    def _recommend(
+        self,
+        searchspace: SearchSpace,
+        batch_quantity: int = 1,
+        allow_repeated_recommendations: bool = False,
+        allow_recommending_already_measured: bool = True,
+    ) -> pd.DataFrame:
+        """
+        Implements the actual recommendation logic. In contrast to its public
+        counterpart, no validation or post-processing is carried out but only the raw
+        recommendation computation is conducted.
+        """
+
+    def check_searchspace_compatibility(self, searchspace: SearchSpace) -> None:
+        """
+        Performs a compatibility check between the recommender type and the provided
+        search space.
+
+        Parameters
+        ----------
+        searchspace : SearchSpace
+            The search space to check for compatibility.
+
+        Raises
+        ------
+        IncompatibleSearchSpaceError
+            In case the recommender is not compatible with the specified search space.
+        """
+        if (not self.compatible_discrete) and (not searchspace.discrete.empty):
+            raise IncompatibleSearchSpaceError(
+                f"The search space that was passed contained a discrete part, but the "
+                f"chosen recommender of type {self.type} does not support discrete "
+                f"search spaces."
+            )
+
+        if (not self.compatible_continuous) and (not searchspace.continuous.empty):
+            raise IncompatibleSearchSpaceError(
+                f"The search space that was passed contained a continuous part, but the"
+                f" chosen recommender of type {self.type} does not support continuous "
+                f"search spaces."
+            )
 
 
 class SequentialGreedyRecommender(Recommender):
@@ -71,12 +141,28 @@ class SequentialGreedyRecommender(Recommender):
         type `botorch.acquisition.monte_carlo.MCAcquisitionFunction`.
     """
 
-    type = "SEQUENTIAL_GREEDY"
+    # TODO: generalize the approach to also support continuous spaces
 
-    def recommend(self, candidates: pd.DataFrame, batch_quantity: int = 1) -> pd.Index:
+    type = "SEQUENTIAL_GREEDY"
+    compatible_discrete = True
+    compatible_continuous = False
+
+    def _recommend(
+        self,
+        searchspace: SearchSpace,
+        batch_quantity: int = 1,
+        allow_repeated_recommendations: bool = False,
+        allow_recommending_already_measured: bool = True,
+    ) -> pd.DataFrame:
         """See base class."""
+
+        candidates_exp, candidates_comp = searchspace.discrete.get_candidates(
+            allow_repeated_recommendations,
+            allow_recommending_already_measured,
+        )
+
         # determine the next set of points to be tested
-        candidates_tensor = to_tensor(candidates)
+        candidates_tensor = to_tensor(candidates_comp)
         try:
             points, _ = optimize_acqf_discrete(
                 self.acquisition_function, batch_quantity, candidates_tensor
@@ -88,24 +174,23 @@ class SequentialGreedyRecommender(Recommender):
             ) from ex
 
         # retrieve the index of the points from the input dataframe
-        # TODO: This additional inelegant step is unfortunately required since BoTorch
-        #   does not return the indices of the points. However, as soon as we move to
-        #   continuous spaces, we will have to use another representation anyway
-        #   (which is also the reason why BoTorch does not support it).
         # IMPROVE: The merging procedure is conceptually similar to what
         #   `SearchSpace._match_measurement_with_searchspace_indices` does, though using
         #   a simpler matching logic. When refactoring the SearchSpace class to
         #   handle continuous parameters, a corresponding utility could be extracted.
-        locs = pd.Index(
+        idxs = pd.Index(
             pd.merge(
-                candidates.reset_index(),
-                pd.DataFrame(points, columns=candidates.columns),
-                on=list(candidates),
+                candidates_comp.reset_index(),
+                pd.DataFrame(points, columns=candidates_comp.columns),
+                on=list(candidates_comp),
             )["index"]
         )
-        assert len(points) == len(locs)
+        assert len(points) == len(idxs)
 
-        return locs
+        rec = candidates_exp.loc[idxs, :]
+        searchspace.discrete.metadata.loc[idxs, "was_recommended"] = True
+
+        return rec
 
 
 class MarginalRankingRecommender(Recommender):
@@ -117,19 +202,36 @@ class MarginalRankingRecommender(Recommender):
     """
 
     type = "UNRESTRICTED_RANKING"
+    compatible_discrete = True
+    compatible_continuous = False
 
-    def recommend(self, candidates: pd.DataFrame, batch_quantity: int = 1) -> pd.Index:
+    def _recommend(
+        self,
+        searchspace: SearchSpace,
+        batch_quantity: int = 1,
+        allow_repeated_recommendations: bool = False,
+        allow_recommending_already_measured: bool = True,
+    ) -> pd.DataFrame:
         """See base class."""
+
+        candidates_exp, candidates_comp = searchspace.discrete.get_candidates(
+            allow_repeated_recommendations,
+            allow_recommending_already_measured,
+        )
+
         # prepare the candidates in t-batches (= parallel marginal evaluation)
-        candidates_tensor = to_tensor(candidates).unsqueeze(1)
+        candidates_tensor = to_tensor(candidates_comp).unsqueeze(1)
 
         # evaluate the acquisition function for each t-batch and construct the ranking
         acqf_values = self.acquisition_function(candidates_tensor)
         ilocs = torch.argsort(acqf_values, descending=True)
 
-        # return the dataframe indices of the top ranked candidates
-        locs = candidates.index[ilocs[:batch_quantity].numpy()]
-        return locs
+        # return top ranked candidates
+        idxs = candidates_comp.index[ilocs[:batch_quantity].numpy()]
+        rec = candidates_exp.loc[idxs, :]
+        searchspace.discrete.metadata.loc[idxs, "was_recommended"] = True
+
+        return rec
 
 
 class RandomRecommender(Recommender):
@@ -138,8 +240,77 @@ class RandomRecommender(Recommender):
     """
 
     type = "RANDOM"
+    compatible_discrete = True
+    compatible_continuous = True
 
-    def recommend(self, candidates: pd.DataFrame, batch_quantity: int = 1) -> pd.Index:
+    def _recommend(
+        self,
+        searchspace: SearchSpace,
+        batch_quantity: int = 1,
+        allow_repeated_recommendations: bool = False,
+        allow_recommending_already_measured: bool = True,
+    ) -> pd.DataFrame:
         """See base class."""
 
-        return candidates.sample(n=batch_quantity).index
+        # Discrete part if applicable
+        rec_disc = pd.DataFrame()
+        if not searchspace.discrete.empty:
+            candidates_exp, _ = searchspace.discrete.get_candidates(
+                allow_repeated_recommendations,
+                allow_recommending_already_measured,
+            )
+
+            # randomly select from discrete candidates
+            rec_disc = candidates_exp.sample(n=batch_quantity)
+            searchspace.discrete.metadata.loc[rec_disc.index, "was_recommended"] = True
+
+        # Continuous part if applicable
+        rec_conti = pd.DataFrame()
+        if not searchspace.continuous.empty:
+            rec_conti = searchspace.continuous.samples_random(n_points=batch_quantity)
+
+        # Merge sub-parts and reorder columns to match original order
+        rec = pd.concat([rec_disc, rec_conti], axis=1).reindex(
+            columns=[p.name for p in searchspace.parameters]
+        )
+
+        return rec
+
+
+# TODO finalize class and type name below
+class PurelyContinuousRecommender(Recommender):
+    """
+    Recommends the next set of experiments by means of sequential greedy optimization
+    in a purely continuous space.
+    """
+
+    type = "CONTI"
+    compatible_discrete = False
+    compatible_continuous = True
+
+    def _recommend(
+        self,
+        searchspace: SearchSpace,
+        batch_quantity: int = 1,
+        allow_repeated_recommendations: bool = False,
+        allow_recommending_already_measured: bool = True,
+    ) -> pd.DataFrame:
+        """See base class."""
+
+        try:
+            points, _ = optimize_acqf(
+                acq_function=self.acquisition_function,
+                bounds=searchspace.param_bounds_comp,
+                q=batch_quantity,
+                num_restarts=5,  # TODO make choice for num_restarts
+                raw_samples=10,  # TODO make choice for raw_samples
+            )
+        except AttributeError as ex:
+            raise NoMCAcquisitionFunctionError(
+                f"The '{self.__class__.__name__}' only works with Monte Carlo "
+                f"acquisition functions."
+            ) from ex
+
+        rec = pd.DataFrame(points, columns=searchspace.continuous.param_names)
+
+        return rec