+++ baybe/searchspace.py
@@ -4,6 +4,7 @@ Functionality for managing search spaces.
 import logging
 from typing import List, Optional, Tuple
 
+import numpy as np
 import pandas as pd
 import torch
 
@@ -17,12 +18,21 @@ from .parameters import (
 from .utils import df_drop_single_value_columns
 
 log = logging.getLogger(__name__)
+INF_BOUNDS_REPLACEMENT = 1000
 
 
 class SearchSpace:
     """
-    Class for managing the overall search space which might be purely discrete, purely
-    continuous or hybrid.
+    Class for managing the overall search space, which might be purely discrete, purely
+    continuous, or hybrid.
+
+    NOTE:
+        Created objects related to the computational representations of parameters
+        (e.g., parameter bounds, computational dataframes, etc.) may use a different
+        parameter order than what is specified through the constructor: While the
+        passed parameter list can contain parameters in arbitrary order, the
+        aforementioned objects (by convention) list discrete parameters first, followed
+        by continuous ones.
     """
 
     def __init__(
@@ -68,13 +78,22 @@ class SearchSpace:
 
     @property
     def contains_mordred(self) -> bool:
-        """Indicates if any of the parameters uses MORDRED encoding."""
-        return any(p.encoding == "MORDRED" for p in self.parameters)
+        """Indicates if any of the discrete parameters uses MORDRED encoding."""
+        return any(p.encoding == "MORDRED" for p in self.discrete.parameters)
 
     @property
     def contains_rdkit(self) -> bool:
-        """Indicates if any of the parameters uses RDKIT encoding."""
-        return any(p.encoding == "RDKIT" for p in self.parameters)
+        """Indicates if any of the discrete parameters uses RDKIT encoding."""
+        return any(p.encoding == "RDKIT" for p in self.discrete.parameters)
+
+    @property
+    def param_bounds_comp(self) -> torch.Tensor:
+        """
+        Returns bounds as tensor.
+        """
+        return torch.hstack(
+            [self.discrete.param_bounds_comp, self.continuous.param_bounds_comp]
+        )
 
     def state_dict(self) -> dict:
         """Creates a dictionary representing the object's internal state."""
@@ -175,6 +194,26 @@ class SubspaceDiscrete:
                 },
                 index=self.exp_rep.index,
             )
+
+    @property
+    def empty(self):
+        """Whether this search space is empty."""
+        return len(self.parameters) == 0
+
+    @property
+    def param_bounds_comp(self) -> torch.Tensor:
+        """
+        Returns bounds as tensor.
+        """
+        if not self.parameters:
+            return torch.empty(2, 0)
+        bounds = np.hstack(
+            [
+                np.vstack([p.comp_df.min().values, p.comp_df.max().values])
+                for p in self.parameters
+            ]
+        )
+        return torch.from_numpy(bounds)
 
     def state_dict(self) -> dict:
         """Creates a dictionary representing the object's internal state."""
@@ -297,7 +336,7 @@ class SubspaceDiscrete:
 
             # We expect exactly one match. If that's not the case, print a warning.
             inds_found = self.exp_rep.index[match].to_list()
-            if len(inds_found) == 0:
+            if len(inds_found) == 0 and len(num_cols) > 0:
                 log.warning(
                     "Input row with index %s could not be matched to the search space. "
                     "This could indicate that something went wrong.",
@@ -318,8 +357,8 @@ class SubspaceDiscrete:
 
     def get_candidates(
         self,
-        allow_repeated_recommendations: bool,
-        allow_recommending_already_measured: bool,
+        allow_repeated_recommendations: bool = False,
+        allow_recommending_already_measured: bool = False,
     ) -> Tuple[pd.DataFrame, pd.DataFrame]:
         """
         Returns the set of candidate parameter settings that can be tested.
@@ -371,7 +410,7 @@ class SubspaceDiscrete:
             A dataframe with the parameters in computational representation.
         """
         # If the transformed values are not required, return an empty dataframe
-        if self.empty_encoding:
+        if self.empty_encoding or len(data) < 1:
             comp_rep = pd.DataFrame(index=data.index)
             return comp_rep
 
@@ -380,7 +419,7 @@ class SubspaceDiscrete:
         for param in self.parameters:
             comp_df = param.transform_rep_exp2comp(data[param.name])
             dfs.append(comp_df)
-        comp_rep = pd.concat(dfs, axis=1)
+        comp_rep = pd.concat(dfs, axis=1) if dfs else pd.DataFrame()
 
         # IMPROVE: The following is a simple mechanism to implement statefulness of
         #   the transformation. However, the state is effectively implemented through
@@ -414,18 +453,25 @@ class SubspaceContinuous:
         self.parameters: List[NumericContinuous] = parameters
 
     @property
-    def bounds(self) -> List[Tuple[float, float]]:
-        """
-        Returns list of parameter bounds.
-        """
-        return [p.bounds for p in self.parameters]
+    def empty(self):
+        """Whether this search space is empty."""
+        return len(self.parameters) == 0
 
     @property
-    def tensor_bounds(self) -> torch.Tensor:
+    def param_names(self) -> List[str]:
+        """
+        Returns list of parameter names.
+        """
+        return [p.name for p in self.parameters]
+
+    @property
+    def param_bounds_comp(self) -> torch.Tensor:
         """
         Returns bounds as tensor.
         """
-        return torch.tensor(self.bounds)
+        if not self.parameters:
+            return torch.empty(2, 0)
+        return torch.tensor([p.bounds for p in self.parameters]).T
 
     def transform(
         self,
@@ -438,3 +484,99 @@ class SubspaceContinuous:
         comp_rep = data[[p.name for p in self.parameters]]
 
         return comp_rep
+
+    # TODO rework the helper functions below, remove finite bound enforcement and
+    #  replace by distributional sampling
+    @property
+    def bounds_forced_finite(self) -> torch.Tensor:
+        """
+        Returns the parameter bounds where infinite values are clipped.
+        """
+        return torch.clip(
+            self.param_bounds_comp, -INF_BOUNDS_REPLACEMENT, INF_BOUNDS_REPLACEMENT
+        )
+
+    @property
+    def is_fully_bounded(self):
+        """
+        Whether the search space has infinite bound or is entirely finitely bounded.
+
+        Returns
+        -------
+        bool
+            True if search space has no infinite bounds.
+        """
+        return torch.isfinite(self.param_bounds_comp)
+
+    def samples_random(self, n_points: int = 1) -> pd.DataFrame:
+        """
+        Get random point samples from the continuous space. Infinite bounds are
+        replaced by half of the maximum floating point number.
+
+        Parameters
+        ----------
+        n_points : int
+            Number of points that should be sampled.
+
+        Returns
+        -------
+        pandas data frame
+            A data frame containing the points as rows with columns corresponding to the
+             parameter names.
+        """
+        if not self.parameters:
+            return pd.DataFrame()
+        points = torch.distributions.uniform.Uniform(*self.bounds_forced_finite).sample(
+            torch.Size((n_points,))
+        )
+        return pd.DataFrame(points, columns=self.param_names)
+
+    def samples_full_factorial(self, n_points: int = 1) -> pd.DataFrame:
+        """
+        Get random point samples from the full factorial of the continuous space.
+
+        Parameters
+        ----------
+        n_points : int
+            Number of points that should be sampled.
+
+        Returns
+        -------
+        pandas data frame
+            A data frame containing the points as rows with columns corresponding to the
+             parameter names.
+        """
+        full_factorial = self.full_factorial
+
+        if len(full_factorial) < n_points:
+            raise ValueError(
+                f"You are trying to sample {n_points} points from the full factorial of"
+                f" the continuous space bounds, but it has only {len(full_factorial)} "
+                f"points."
+            )
+
+        return full_factorial.sample(n=n_points).reset_index(drop=True)
+
+    @property
+    def full_factorial(self) -> pd.DataFrame:
+        """
+        Get the full factorial of the continuous space.
+
+        Returns
+        -------
+        pandas data frame
+            A data frame containing the full factorial
+        """
+        if not self.is_fully_bounded:
+            log.warning(
+                "You are trying to access the full factorial of a continuous sace that "
+                "has infinite bounds in at least one parameter. Internally, infinite "
+                "bounds have been replaced by -/+ %f",
+                INF_BOUNDS_REPLACEMENT,
+            )
+
+        index = pd.MultiIndex.from_product(
+            self.bounds_forced_finite.T.tolist(), names=self.param_names
+        )
+
+        return pd.DataFrame(index=index).reset_index()